sortflow_web_service/requirements.txt
# requirements.txt - SortFlow Web Service

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

werkzeug==3.0.3
bcrypt==4.2.0
PyJWT==2.9.0

SQLAlchemy==2.0.35

requests==2.32.3

python-dotenv==1.0.1